requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
cx-Oracle==8.3.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-import-export==2.7.1
et-xmlfile==1.1.0
iniconfig==1.1.1
MarkupPy==1.14
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pymongo==4.0.1
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==6.2.5
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
toml==0.10.2
tzdata==2021.5
xlrd==2.0.1
xlwt==1.3.0
